Change array::size_type from uint64 to uint32

+++ engine/include/RedEngine/Rendering/Renderer.hpp
@@ -90,7 +90,7 @@ public:
 
 private:
     Array<RenderingData>& GetVisibleRenderDatasForType(RenderEntityType type, CameraComponent* camera,
-                                                       uint64& renderDataCount);
+                                                       uint32& renderDataCount);
 
     void UseMaterial(const MaterialInstance& mat);
     void UseGeometry(const Geometry* geom);
